requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langgraph
google-generativeai
beautifulsoup4
python-dotenv
requests
sentence-transformers
numpy